Improved GraphQL client
This is a complete rewrite of the GraphQL client functionality.

The previous implementation suffered from a few problems making it
virtually unusable as detailed in
https://github.com/Shopify/shopify_api/issues/511. Here's a summary:

1. you couldn't specify a local schema file
2. due to the above, the client made a dynamic introspection request on
initialization which was very slow
3. unbounded memory growth due to building new clients at runtime

This rewrite was focused on solving those problems first and foremost
but we also had a few other goals:

* support API versioning
* provide better defaults and an improved developer experience
* ensure it's impossible to do the wrong thing

The new GraphQL client *only* supports loading local schema files to
ensure no introspection requests are made during app runtime. The goal
is that clients are fully initialized at application boot time (and if
you're using Rails this is handled automatically).

Workflow:
1. Set `ShopifyAPI::GraphQL.schema_location` to a directory path (or
use the default in Rails of `db/shopify_graphql_schemas`).
2. Save a JSON version of Shopify's Admin schema locally (or use the
`shopify_api:graphql:dump` Rake task) to the `schema_location` and name
it after the API version: `2020-01.json`.
3. Access the client at `ShopifyAPI::GraphQL.client`
4. Execute queries via `client.query`